+++ stock/views/temp/create.py
@@ -1,5 +1,6 @@
 from stock.models import TempExpense
 from django.utils import timezone
+from django.contrib.auth.models import User
 
 
 def create(request):
@@ -21,10 +22,20 @@ def edit(request):
     if int(value) < 0:
         return 'Amount is not valid'
     if 'DELETE' in request.POST:
-        TempExpense.objects.get(id=id).delete()
-        return 'deleted'
+        if check_password(request):
+            TempExpense.objects.get(id=id).delete()
+            return 'deleted'
+        return 'password is not correct'
     else:
         update = TempExpense.objects.get(id=id)
         update.value = int(value)
         update.save()
         return 'updated'
+
+
+def check_password(request):
+
+    if User.objects.get(username=request.user). \
+            check_password(request.POST.get('form_verifypwd')):
+        return True
+    return False
